BUG: month and quarter offsets truncated n to 32 bits (GH#66549)
`n` was carried through the month-arithmetic layer in a C `int`, so a
multiple of 2**32 silently became a no-op and n >= 2**31 wrapped negative,
on both the scalar and the vectorized paths.

Widen the period count to int64_t in year_add_months, month_add_months,
shift_months, shift_quarters, shift_month, roll_convention, _roll_qtrday
and SemiMonthOffset._apply_array, and widen the year argument of
get_days_in_month, get_lastbday and get_firstbday, which otherwise decide
leap years from a wrapped year at the magnitudes this newly makes reachable.

+++ pandas/_libs/tslibs/ccalendar.pyx
@@ -52,13 +52,13 @@ weekday_to_int = {int_to_weekday[key]: key for key in int_to_weekday}
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
-cpdef int32_t get_days_in_month(int year, Py_ssize_t month) noexcept nogil:
+cpdef int32_t get_days_in_month(int64_t year, Py_ssize_t month) noexcept nogil:
     """
     Return the number of days in the given month of the given year.
 
     Parameters
     ----------
-    year : int
+    year : int64_t
     month : int
 
     Returns
@@ -264,13 +264,13 @@ cpdef int32_t get_day_of_year(int year, int month, int day) noexcept nogil:
 # ---------------------------------------------------------------------
 # Business Helpers
 
-cpdef int get_lastbday(int year, int month) noexcept nogil:
+cpdef int get_lastbday(int64_t year, int month) noexcept nogil:
     """
     Find the last day of the month that is a business day.
 
     Parameters
     ----------
-    year : int
+    year : int64_t
     month : int
 
     Returns
@@ -285,13 +285,13 @@ cpdef int get_lastbday(int year, int month) noexcept nogil:
     return days_in_month - max(((wkday + days_in_month - 1) % 7) - 4, 0)
 
 
-cpdef int get_firstbday(int year, int month) noexcept nogil:
+cpdef int get_firstbday(int64_t year, int month) noexcept nogil:
     """
     Find the first day of the month that is a business day.
 
     Parameters
     ----------
-    year : int
+    year : int64_t
     month : int
 
     Returns
